fix1

+++ on-complete.sh
@@ -63,6 +63,9 @@ File path: ${FILE_PATH}
 Upload path: ${UPLOAD_PATH}
 Remote path A: ${REMOTE_PATH}
 Remote path B: ${REMOTE_PATH_2}
+Remote path C: ${REMOTE_PATH_3}
+Remote path D: ${REMOTE_PATH_4}
+Remote path E: ${REMOTE_PATH_5}
 -------------------------- [TASK INFO] --------------------------
 "
 }
@@ -86,11 +89,26 @@ UPLOAD_FILE() {
         rclone copy -v "${UPLOAD_PATH}" "${REMOTE_PATH}"
         RCLONE_EXIT_CODE=$?
 		RCLONE_EXIT_CODE_2=0
+		RCLONE_EXIT_CODE_3=0
+		RCLONE_EXIT_CODE_4=0
+		RCLONE_EXIT_CODE_5=0
 		if [ -n "${RCLONE_DESTINATION_2}" ]; then
 			rclone copy -v "${UPLOAD_PATH}" "${REMOTE_PATH_2}"
 			RCLONE_EXIT_CODE_2=$?
 		fi
-        if [ ${RCLONE_EXIT_CODE} -eq 0 ] && [ ${RCLONE_EXIT_CODE_2} -eq 0 ]; then
+		if [ -n "${RCLONE_DESTINATION_3}" ]; then
+			rclone copy -v "${UPLOAD_PATH}" "${REMOTE_PATH_2}"
+			RCLONE_EXIT_CODE_2=$?
+		fi
+		if [ -n "${RCLONE_DESTINATION_4}" ]; then
+			rclone copy -v "${UPLOAD_PATH}" "${REMOTE_PATH_2}"
+			RCLONE_EXIT_CODE_2=$?
+		fi
+		if [ -n "${RCLONE_DESTINATION_5}" ]; then
+			rclone copy -v "${UPLOAD_PATH}" "${REMOTE_PATH_2}"
+			RCLONE_EXIT_CODE_2=$?
+		fi
+        if [ ${RCLONE_EXIT_CODE} -eq 0 ] && [ ${RCLONE_EXIT_CODE_2} -eq 0 ] && [ ${RCLONE_EXIT_CODE_3} -eq 0 ] && [ ${RCLONE_EXIT_CODE_4} -eq 0 ] && [ ${RCLONE_EXIT_CODE_5} -eq 0 ]; then
             [ -e "${DOT_ARIA2_FILE}" ] && rm -vf "${DOT_ARIA2_FILE}"
             rclone rmdirs -v "${DOWNLOAD_PATH}" --leave-root
             echo -e "$(date +"%m/%d %H:%M:%S") ${INFO} Upload done: ${UPLOAD_PATH}"
@@ -133,6 +151,9 @@ if [ "${TOP_PATH}" = "${FILE_PATH}" ] && [ $2 -eq 1 ]; then # 普通单文件下
     UPLOAD_PATH="${FILE_PATH}"
     REMOTE_PATH="${RCLONE_DESTINATION}/"
     REMOTE_PATH_2="${RCLONE_DESTINATION_2}/"
+    REMOTE_PATH_2="${RCLONE_DESTINATION_3}/"
+    REMOTE_PATH_2="${RCLONE_DESTINATION_4}/"
+    REMOTE_PATH_2="${RCLONE_DESTINATION_5}/"
     UPLOAD
     exit 0
 elif [ "${TOP_PATH}" != "${FILE_PATH}" ] && [ $2 -gt 1 ]; then # BT下载（文件夹内文件数大于1），移动整个文件夹到设定的网盘文件夹。
